add four more variables
